Cambio contratos controllers y useCases

+++ api/src/main/java/com/coding/challenge/api/transactionEndpoints/dtos/responses/GetTransactionSumResponse.java
@@ -1,10 +1,12 @@
 package com.coding.challenge.api.transactionEndpoints.dtos.responses;
 
 
+import java.math.BigDecimal;
+
 import io.swagger.v3.oas.annotations.media.Schema;
 
 @Schema(name = "SumResponse", description = "Sum of the amounts for a transaction graph")
 public record GetTransactionSumResponse(
     @Schema(description = "Aggregate sum (root + all descendants)", example = "7500.5")
-    double sum
+    BigDecimal sum
 ) {}
